dll: sync DragDropManager, set :dragging pseudo-state, fix macOS type mismatches
- Sync DragDropManager.active_drag from GestureAndDragManager after drag activation
- Set :dragging pseudo-state on source node at DragStart
- Clear :dragging and :drag-over on DragEnd
- Fix maximumFramesPerSecond return type from f64 to i64 (NSInteger)
- Fix NSOpenGLContextParameter type from i32 to i64 (NSInteger)

+++ dll/src/desktop/display.rs
@@ -501,14 +501,14 @@ mod macos {
                 i
             );
             let refresh_rate = unsafe {
-                let fps: f64 = msg_send![&**screen, maximumFramesPerSecond];
+                let fps: i64 = msg_send![&**screen, maximumFramesPerSecond];
                 log_debug!(
                     LogCategory::General,
                     "[get_displays] Screen {} refresh rate: {} Hz",
                     i,
                     fps
                 );
-                if fps > 0.0 {
+                if fps > 0 {
                     fps as u16
                 } else {
                     60
